bot/requirements.txt
# Discord Bot Dependencies
# Install with: pip install -r requirements-discord.txt

discord.py>=2.3.2
python-dotenv>=1.0.0
